TwoFace refactoring
