+++ version.sbt
@@ -1,1 +1,1 @@
-version in ThisBuild := "0.0.5-SNAPSHOT"
+version in ThisBuild := "0.1.3-SNAPSHOT"
